requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.2
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyppeteer==2.0.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.29.0
websockets==10.4
zipp==3.18.1
